Feat: Readme.md 이미지 및 비밀번호 찾기 페이지 로직수정
