+++ package.json
@@ -38,7 +38,6 @@
         "react-dom": "^18.3.1",
         "react-hook-form": "^7.53.0",
         "react-icons": "^5.3.0",
-        "react-quill": "^2.0.0",
         "react-router-dom": "^6.26.0",
         "tailwind-merge": "^2.5.2",
         "zustand": "^4.5.4"
