+++ src/api/axios.ts
@@ -2,9 +2,9 @@ import axios from "axios";
 
 export const axiosInstance = axios.create({
     //? LOCAL
-    // baseURL: "/api",
+    baseURL: "/api",
     //? SERVER
-    baseURL: "https://api.hunsuking.yoyobar.xyz/api",
+    // baseURL: "https://api.hunsuking.yoyobar.xyz/api",
     headers: {
         "Content-Type": "application/json",
         "Cache-Control": "no-cache",
